Added Branch and Leaf prefabs
Added Branch and Leaf Prefabs, altered movement to enable left & right movement on the ground and in air, we should talk about whether we want that. I found as I was implementing my branches that it was very difficult to control much of anything without ground movement and in-air movement.
I also edited the temp player sprite to have a directional arrow, which makes it easier to identify which way the sprite is going to go.